Fix ONNX export

+++ utils/export_damoyolo.py
@@ -18,7 +18,7 @@ class DeepStreamOutput(nn.Module):
     def forward(self, x):
         boxes = x[1]
         scores, classes = torch.max(x[0], 2, keepdim=True)
-        return torch.cat((boxes, scores, classes), dim=2)
+        return torch.cat((boxes, scores, classes.float()), dim=2)
 
 
 def suppress_warnings():
